Restrict creation of security properties if OIDC is enabled
