+++ deduplication/src/test/java/gov/cdc/nbs/deduplication/auth/LoginControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import gov.cdc.nbs.deduplication.auth.model.LoginRequest;
 import gov.cdc.nbs.deduplication.auth.model.LoginResponse;
-import gov.cdc.nbs.deduplication.config.auth.AuthenticationConfiguration.SecurityProperties;
+import gov.cdc.nbs.deduplication.config.auth.nbs.NbsTokenConfiguration.SecurityProperties;
 import gov.cdc.nbs.deduplication.config.auth.nbs.token.NbsToken;
 import gov.cdc.nbs.deduplication.config.auth.nbs.token.NbsTokenCreator;
 import gov.cdc.nbs.deduplication.config.auth.user.NbsUserDetails;
